# web-ui/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AIMS Web Application API
  description: |
    Complete API specification for the AIMS (Internet Media Store) Web Application.
    
    This API provides comprehensive e-commerce functionality including:
    - Product catalog management
    - Shopping cart operations
    - Order processing and tracking
    - Payment integration with VNPay
    - User authentication and management
    - Administrative operations
    
    ## Authentication
    This API uses JWT Bearer token authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer your_jwt_token_here
    ```
    
    ## Rate Limiting
    API requests are rate limited per user:
    - General endpoints: 1000 requests/hour
    - Authentication: 10 requests/minute
    - File uploads: 100 requests/hour
    
  version: 1.0.0
  contact:
    name: AIMS Development Team
    email: dev@aims.com
  license:
    name: Proprietary
    url: https://aims.com/license
servers:
  - url: https://api.aims.com
    description: Production server
  - url: https://staging-api.aims.com
    description: Staging server
  - url: http://localhost:8080/api
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /products/{productId}:
    get:
      tags:
        - Products
      summary: Get product by ID
      description: Retrieve detailed information about a specific product
      security: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Cart Endpoints
...
  /admin/products:
    get:
      tags:
        - Admin
      summary: Get products for admin
      description: Retrieve products with admin-specific information
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Admin products retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminProductListResponse'

    post:
      tags:
        - Admin
      summary: Create new product
      description: Create a new product in the catalog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: string
          example: prod_123
        title:
          type: string
          example: "The Great Gatsby"
        price:
          type: number
          format: float
          example: 15.99
        category:
          type: string
          example: "Classic Literature"
        productType:
          type: string
          enum: [Book, CD, DVD, LP]
          example: Book
        author:
          type: string
          example: "F. Scott Fitzgerald"
        description:
          type: string
          example: "A classic American novel..."
        imageUrl:
          type: string
          example: "https://images.aims.com/products/great-gatsby.jpg"
        stockQuantity:
          type: integer
          example: 25
        isAvailable:
          type: boolean
          example: true
        specifications:
          type: object
          additionalProperties: true
...
    ProductResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Product retrieved successfully
        data:
          $ref: '#/components/schemas/Product'

    # Cart Schemas
...
    AdminProductListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Admin products retrieved successfully
        data:
          type: array
          items:
            allOf:
              - $ref: '#/components/schemas/Product'
              - type: object
                properties:
                  salesCount:
                    type: integer
                    example: 45
                  revenue:
                    type: number
                    format: float
                    example: 719.55
        pagination:
          $ref: '#/components/schemas/Pagination'
    CreateProductRequest:
      type: object
      required:
        - title
        - price
        - category
        - productType
      properties:
        title:
          type: string
          example: "New Product Title"
        price:
          type: number
          format: float
          example: 19.99
        category:
          type: string
          example: "Technology"
        productType:
          type: string
          enum: [Book, CD, DVD, LP]
          example: Book
        description:
          type: string
          example: "Product description..."
        author:
          type: string
          example: "Author Name"
        stockQuantity:
          type: integer
          example: 100

    # Common Schemas
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 20
        total:
          type: integer
          example: 150
        totalPages:
          type: integer
          example: 8
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: An error occurred
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: email
              code:
                type: string
                example: INVALID_EMAIL
              message:
                type: string
                example: Please provide a valid email address
...